spaces/requirements.txt
torch>=1.7
torchvision>=0.8
matplotlib>=3.0
pandas>=1.0
numpy>=1.19
Pillow>=8.0
scikit-learn>=0.24
requests>=2.20
gradio>=4.0
